copied socket code to socket-server.js
